requirements.txt
fastapi==0.104.1
google-generativeai==0.3.2
pydantic==2.5.0